CrudCar Update
Started to work on Crud Car. Added the fields for the class and some methods. Both the and the column in the database must be renamed: transmission (they are transimission). Also renamed the fields of Car dao: engine and transmission to: engineId and transmissionId and set their type to int.